Added solid red line after the Contact text.
Applied negative margin to the largePinkEllipseColumn

Fixed the margin on the Interested? Let's chat

Fixed padding on all the green, blue and pink sections. Different padding for different text areas for the entire page.

Redid the greenBlueGradient from photoshop.

Added negative margin-top to the pink ellipse

Added the project to my server.